regression: UserInfoTab Broken (#22019)
