+++ client/views/room/contextualBar/UserInfo/UserInfo.js
@@ -5,6 +5,7 @@ import MarkdownText from '../../../../components/MarkdownText';
 import UTCClock from '../../../../components/UTCClock';
 import UserCard from '../../../../components/UserCard';
 import VerticalBar from '../../../../components/VerticalBar';
+import { useSetting } from '../../../../contexts/SettingsContext';
 import { useTranslation } from '../../../../contexts/TranslationContext';
 import { useTimeAgo } from '../../../../hooks/useTimeAgo';
 import InfoPanel from '../../../InfoPanel';
@@ -33,6 +34,19 @@ function UserInfo({
 	const t = useTranslation();
 
 	const timeAgo = useTimeAgo();
+
+	const customFieldsToShowSetting = useSetting('Accounts_CustomFieldsToShowInUserInfo');
+	const customFieldsToShowObj = JSON.parse(customFieldsToShowSetting);
+
+	const customFieldsToShow = customFieldsToShowObj
+		? Object.values(customFieldsToShowObj).map((value) => {
+				const role = Object.values(value);
+				const roleNameToShow = Object.keys(value);
+				const customField = {};
+				customField[roleNameToShow] = customFields[role];
+				return customField;
+		  })
+		: [];
 
 	return (
 		<VerticalBar.ScrollableContent p='x24' {...props}>
@@ -129,7 +143,7 @@ function UserInfo({
 						</InfoPanel.Field>
 					)}
 
-					{customFields.map((customField) =>
+					{customFieldsToShow.map((customField) =>
 						Object.values(customField)[0] ? (
 							<InfoPanel.Field key={Object.keys(customField)[0]}>
 								<InfoPanel.Label>{t(Object.keys(customField)[0])}</InfoPanel.Label>
